Split apart tests
