+++ src/server/__tests__/db.test.ts
@@ -2,29 +2,10 @@ import { getRepository } from "typeorm";
 
 import { dbConnection } from "../db";
 import { User } from "../entity/User";
-import { genAccessTokenInfo, getLoginURL } from "../google";
 
 beforeAll(
   async (): Promise<void> => {
     await dbConnection();
-  },
-);
-
-test(
-  "getLoginURL",
-  async (): Promise<void> => {
-    const url: string = getLoginURL();
-    expect(url)
-      .toMatch("/accounts.google.com/");
-  },
-);
-
-test(
-  "genAccessTokenInfo",
-  async (): Promise<void> => {
-    await expect(genAccessTokenInfo("ERROR"))
-      .rejects
-      .toThrow(Error);
   },
 );
