+++ src/server/__tests__/google.test.ts
@@ -1,0 +1,19 @@
+import { genAccessTokenInfo, getLoginURL } from "../google";
+
+test(
+  "getLoginURL",
+  async (): Promise<void> => {
+    const url: string = getLoginURL();
+    expect(url)
+      .toMatch("/accounts.google.com/");
+  },
+);
+
+test(
+  "genAccessTokenInfo",
+  async (): Promise<void> => {
+    await expect(genAccessTokenInfo("ERROR"))
+      .rejects
+      .toThrow(Error);
+  },
+);
